# Optimize MapOf.Range