Before:
```
BenchmarkMapOfRange-8   	 1000000	      1283 ns/op	    779361 ops/s	       0 B/op	       0 allocs/op
```

After:
```
BenchmarkMapOfRange-8   	 1241162	       961.4 ns/op	   1040141 ops/s	       0 B/op	       0 allocs/op
```